requirements.txt
python-telegram-bot==21.0
fastapi
uvicorn[standard]
